Wamya-backend-adapters/Wamya-backend-persistence/src/test/java/com/excentria_it/wamya/adapter/persistence/mapper/UserPreferenceMapperTests.java
package com.excentria_it.wamya.adapter.persistence.mapper;

import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.Test;

import com.excentria_it.wamya.adapter.persistence.entity.UserPreferenceJpaEntity;
import com.excentria_it.wamya.domain.UserPreference;

public class UserPreferenceMapperTests {
	private UserPreferenceMapper userPreferenceMapper = new UserPreferenceMapper();

	@Test
	void testMapToDomainEntityWithNullJpaEntity() {
		assertNull(userPreferenceMapper.mapToDomainEntity(null));
	}

	@Test
	void testMapToDomainEntity() {
		UserPreferenceJpaEntity jpaEntity = new UserPreferenceJpaEntity("timezone", "Africa/Tunis");
		jpaEntity.setId(1L);

		UserPreference userPreference = userPreferenceMapper.mapToDomainEntity(jpaEntity);

		assertEquals(jpaEntity.getId(), userPreference.getId());
		assertEquals(jpaEntity.getKey(), userPreference.getKey());
		assertEquals(jpaEntity.getValue(), userPreference.getValue());
	}
}
